requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-HTTPAuth==4.8.0
requests==2.31.0
requests-cache==1.1.1
python-dotenv==1.0.0